Training calc tweaks
I felt that the training option was too expensive to be meaningful at any level, given that there is already a 180G buy-in. I've tweaked it to scale based on the Level EXP requirements set in the world settings to make it more useful. In keeping with the original vision of diminishing returns, the EXP given by each level is a now linear function, whereas the cost remains an exponential function. From my cursory check, the gap in effectiveness shouldn't be too bad unless crazy buildings are enabled (which are already stated to be unbalanced).